Remove recursive seq_view usage

+++ stan/math/prim/fun/num_elements.hpp
@@ -16,7 +16,7 @@ namespace math {
  * @return 1
  */
 template <typename T, require_stan_scalar_t<T>* = nullptr>
-inline int num_elements(const T& x) {
+inline size_t num_elements(const T& x) {
   return 1;
 }
 
@@ -29,7 +29,7 @@ inline int num_elements(const T& x) {
  * @return size of matrix
  */
 template <typename T, require_matrix_t<T>* = nullptr>
-inline int num_elements(const T& m) {
+inline size_t num_elements(const T& m) {
   return m.size();
 }
 
@@ -43,7 +43,7 @@ inline int num_elements(const T& m) {
  * @return number of contained arguments
  */
 template <typename T>
-inline int num_elements(const std::vector<T>& v) {
+inline size_t num_elements(const std::vector<T>& v) {
   if (v.size() == 0) {
     return 0;
   }
